+++ CHANGELOG.md
@@ -111,6 +111,9 @@ to avoid letting the user edit the immutable one.
 * `jj rebase -r` now accepts `--insert-after` and `--insert-before` options to
   customize the location of the rebased revisions.
 
+* A new revset `reahable(srcs, domain)` will return all commits that are
+  reachable from `srcs` within `domain`.
+
 ### Fixed bugs
 
 * Revsets now support `\`-escapes in string literal.